Move iterator code into a module
